Fixed http log query logic + some small fixes here and there
The query for HTTP log counts was incorrect, using a SUM() of statements
counts in batches. That is unreliable as the same batch could have
multiple http requests under certain circumstances. A count of all items
for a certain code inthe log is now uses instead. There were a few typos
and bugs that did not affect the functionality all that much. Added the
long forgotten composer.json file as it now does point to the correct
URLs.