Refactoring and admin menus.
Introduced a method that would just send the statements data to LRS.
That method is noe used by both connection test and sending data to LRS
logic. Changed connection testing code to use the current values within
the form and not the already stored ones as that was confusing. Changed
the path to a module configuration page and included that into the admin
menu. Also registered the statistics page as a submenu for the Reports.